# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IPN-DEV — Sistema de Inventario Daluzed Pastelería",
        "description": "API REST para gestión de inventario, producción y logística de Daluzed Pastelería. Proyecto Nuclear 3 · Corporación Universitaria Alexander von Humboldt (CUE).",
        "contact": {
            "name": "Equipo IPN-DEV",
            "email": "admin@inventario.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/recepciones/ordenes": {
            "get": {
                "tags": [
                    "Recepciones"
                ],
                "summary": "Listar órdenes de pedido",
                "operationId": "8f6a4ba374169eb8539d234a771f3d41",
                "responses": {
                    "200": {
                        "description": "Listado de órdenes"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso (requiere recepciones.leer)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recepciones"
                ],
                "summary": "Crear orden de pedido",
                "operationId": "809c144dfb8d95d8c635fb6a0583c69a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "proveedor"
                                ],
                                "properties": {
                                    "proveedor": {
                                        "type": "string",
                                        "example": "Harinera del Valle"
                                    },
                                    "fecha_esperada": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-15"
                                    },
                                    "observaciones": {
                                        "type": "string",
                                        "example": "Entrega urgente"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso (requiere recepciones.escribir)"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/recepciones/ordenes/{id}": {
            "get": {
                "tags": [
                    "Recepciones"
                ],
                "summary": "Ver orden de pedido",
                "operationId": "b49577e1aa8b72f280132d81018a02d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la orden"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Recepciones"
                ],
                "summary": "Actualizar orden de pedido",
                "operationId": "65c31af806ed75d5f438090b0e3b2920",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "proveedor": {
                                        "type": "string"
                                    },
                                    "fecha_esperada": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "en_recepcion",
                                            "cerrada"
                                        ]
                                    },
                                    "observaciones": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Orden actualizada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso"
                    },
                    "404": {
                        "description": "No encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/recepciones": {
            "get": {
                "tags": [
                    "Recepciones"
                ],
                "summary": "Listar recepciones",
                "operationId": "25f26a7fe9f45c488b0be4883ffc2d08",
                "responses": {
                    "200": {
                        "description": "Listado de recepciones"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/recepciones/ordenes/{id}/recepciones": {
            "post": {
                "tags": [
                    "Recepciones"
                ],
                "summary": "Registrar recepción de materias primas",
                "description": "Registra la entrada física de MP contra una orden de pedido. Crea lotes con trazabilidad FEFO (RFREC).",
                "operationId": "b7c2588158024293a8211fa7c029c6ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la orden de pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "materia_prima_id",
                                                "cantidad"
                                            ],
                                            "properties": {
                                                "materia_prima_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "cantidad": {
                                                    "type": "number",
                                                    "example": 100
                                                },
                                                "fecha_vencimiento": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2026-12-31"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recepción registrada, lotes creados en Bodega Principal"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso (requiere recepciones.escribir)"
                    },
                    "404": {
                        "description": "Orden no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/recepciones/{id}": {
            "get": {
                "tags": [
                    "Recepciones"
                ],
                "summary": "Ver recepción",
                "operationId": "0b522b450c02bee9ed10a61624cc7e38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la recepción con lotes creados"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Bearer generado por Laravel Sanctum al hacer login. Incluirlo en el header: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}